--- a/src/DynamicMvvm.Benchmarks/Program.cs
+++ b/src/DynamicMvvm.Benchmarks/Program.cs
@@ -1,0 +1,4 @@
+﻿using BenchmarkDotNet.Running;
+using DynamicMvvm.Benchmarks;
+
+BenchmarkRunner.Run<Benchmark>();