… BREAKING_CHANGES.md
… src/DynamicMvvm/ViewModel/ViewModelBase.Disposable.cs
… src/DynamicMvvm/ViewModel/ViewModelBase.Errors.cs
… src/DynamicMvvm/ViewModel/ViewModelBase.PropertyChanged.cs