GitHub Issue: #
<!-- Link to relevant GitHub issue if applicable.
     All PRs should be associated with an issue -->

## Proposed Changes
<!-- Please check one or more that apply to this PR. -->

 - [ ] Bug fix
 - [x] Feature
 - [ ] Code style update (formatting)
 - [x] Refactoring (no functional changes, no api changes)
 - [ ] Build or CI related changes
 - [ ] Documentation content changes
 - [ ] Other, please describe:

## Description

- Add a benchmark project to measure the performance of some common scenarios.
- All logs now use `LoggerMessage` to reduce memory allocations and speed-up the execution when the log levels are disabled.

## What is the current behavior?
<!-- Please describe the current behavior that you are modifying,
     or link to a relevant issue. -->

| Method                              | Mean      | Allocated |
|------------------------------------ |----------:|----------:|
| CreateViewModel                     |  2.278 μs |    2023 B |
| CreateViewModel_WithExplicitName    |           |           |
| ReadProperty_Unresolved             |  2.754 μs |    1032 B |
| ReadProperty_Resolved               |  1.409 μs |     896 B |
| SetProperty_Unresolved              | 19.147 μs |    3016 B |
| SetProperty_Resolved                |  1.957 μs |     832 B |
| DisposeViewModel                    |  2.954 μs |     776 B |
| SetProperty_Unresolved_WithListener | 21.014 μs |    3256 B |
| SetProperty_Resolved_WithListener   |  3.031 μs |    1088 B |
| DisposeViewModel_WithListener       |  2.932 μs |     776 B |

## What is the new behavior?
<!-- Please describe the new behavior after your modifications. -->

| Method                              | Mean      | Allocated |
|------------------------------------ |----------:|----------:|
| CreateViewModel                     |  2.057 μs |    1760 B |
| CreateViewModel_WithExplicitName    |  2.077 μs |    1760 B |
| ReadProperty_Unresolved             |  2.866 μs |     856 B |
| ReadProperty_Resolved               |  1.423 μs |     720 B |
| SetProperty_Unresolved              | 22.519 μs |    2592 B |
| SetProperty_Resolved                |  1.837 μs |     656 B |
| DisposeViewModel                    |  1.340 μs |     600 B |
| SetProperty_Unresolved_WithListener | 21.484 μs |    2616 B |
| SetProperty_Resolved_WithListener   |  2.303 μs |     680 B |
| DisposeViewModel_WithListener       |  1.301 μs |     600 B |

## Impact on version
<!-- Please select one or more based on your commits. -->

- [ ] **Major** (Public API was modified.)
  - Public constructs (class, struct, delegate, enum, etc.) were removed or renamed.
  - Public members were removed or renamed.
  - Public method signatures were changed or renamed.
- [ ] **Minor** (Public API was extended.)
  - Public constructs, members, or overloads were added.
- [x] **Patch** (Public API was unchanged.)
  - A bug in behavior was fixed.
  - Documentation was changed.
- [ ] **None** (The library is unchanged.)
  - Only code under the `build` folder was changed.
  - Only code under the `.github` folder was changed.

## Checklist

Please check that your PR fulfills the following requirements:

- [ ] Documentation has been added/updated.
- [ ] Automated Unit / Integration tests for the changes have been added/updated.
- [x] Updated [BREAKING_CHANGES.md](../BREAKING_CHANGES.md) (if you introduced a breaking change).
- [x] Your conventional commits are aligned with the **Impact on version** section.
  - The version bump is minor even though this is a patch because it contains a minor breaking change (bump of nuget dependency versions) .

<!-- If this PR contains a breaking change, please describe the impact
     and migration path for existing applications below. -->

## Other information
<!-- Please provide any additional information if necessary -->